made sure changes are commited

+++ src/BankPackage/Account.java
@@ -3,10 +3,10 @@ package BankPackage;
 public abstract class Account {
 
 	//Withdraw
-	public abstract void withdraw(int amount);
+	public abstract void withdraw(float amount);
 	
 	//Deposit
-	public abstract void deposit(int amount);
+	public abstract void deposit(float amount);
 	
 	//Check  amount in Balance or CurrentBalance
 	public abstract void check();
